+++ cmd/grafana-multi-tenant-proxy/main.go
@@ -13,26 +13,29 @@ var (
 
 func main() {
 	app := cli.NewApp()
-	app.Name = "Loki Multitenant Proxy"
-	app.Usage = "Makes your Loki server multi tenant"
+	app.Name = "Grafana Multitenant Proxy"
+	app.Usage = "Makes Grafana Labs applications multi tenant"
 	app.Version = version
 	app.Authors = []*cli.Author{
 		{Name: "Angel Barrera", Email: "angel@k8spin.cloud"},
 		{Name: "Pau Rosello", Email: "pau@k8spin.cloud"},
+		{Name: "Herve Nicol", Email: "herve@giantswarm.io"},
+		{Name: "Quentin Bisson", Email: "quentin@giantswarm.io"},
+		{Name: "Marie Roque", Email: "marie@giantswarm.io"},
 	}
 	app.Commands = []*cli.Command{
 		{
 			Name:   "run",
-			Usage:  "Runs the Loki multi tenant proxy",
+			Usage:  "Runs the Grafana multi tenant proxy",
 			Action: proxy.Serve,
 			Flags: []cli.Flag{
 				&cli.IntFlag{
 					Name:  "port",
-					Usage: "Port to expose this loki proxy",
+					Usage: "Port to expose this proxy",
 					Value: 3501,
 				}, &cli.StringFlag{
-					Name:  "loki-server",
-					Usage: "Loki server endpoint",
+					Name:  "target-server",
+					Usage: "Target server endpoint",
 					Value: "http://localhost:3500",
 				}, &cli.StringFlag{
 					Name:  "auth-config",