+++ internal/app/grafana-multi-tenant-proxy/reverse.go
@@ -5,8 +5,8 @@ import (
 	"net/http/httputil"
 )
 
-// ReverseLoki reverse proxies to Loki
-func ReverseLoki(reverseProxy *httputil.ReverseProxy) http.HandlerFunc {
+// ReverseTarget reverse proxies to target server
+func ReverseTarget(reverseProxy *httputil.ReverseProxy) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		reverseProxy.ServeHTTP(w, r)
 	}